rename: src/constant -> src/constants
