+++ src/features/diaryList/components/DiaryCard.css.ts
@@ -1,5 +1,5 @@
 import { style } from '@vanilla-extract/css';
-import { COLORS } from '@/constant/colors.ts';
+import { COLORS } from '@/constants/colors';
 
 export const li = style({
   width: '100%',
