+++ src/features/diaryList/components/DiaryListHeader.css.ts
@@ -1,5 +1,5 @@
 import { style } from '@vanilla-extract/css';
-import { COLORS } from '@/constant/colors.ts';
+import { COLORS } from '@/constants/colors';
 
 export const listHeader = style({
   display: 'flex',
